+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.24.0] - 2025-02-05
+
 ### Added
 
 - Added `runtime.txt` support for the `python-3.X` major Python version form. ([#322](https://github.com/heroku/buildpacks-python/pull/322))
@@ -296,7 +298,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation. ([#3](https://github.com/heroku/buildpacks-python/pull/3))
 
-[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.23.0...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.24.0...HEAD
+[0.24.0]: https://github.com/heroku/buildpacks-python/compare/v0.23.0...v0.24.0
 [0.23.0]: https://github.com/heroku/buildpacks-python/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/heroku/buildpacks-python/compare/v0.21.0...v0.22.0
 [0.21.0]: https://github.com/heroku/buildpacks-python/compare/v0.20.1...v0.21.0